+++ frontend/src/1-app/routes/Router.jsx
@@ -1,9 +1,9 @@
 import { Routes, Route } from "react-router-dom";
-import Home from "../../pages/home/Home.jsx";
-import PrivacyPolicy from "../../pages/legal-documents/privacy-policy/Privacy-Policy.jsx";
-import Terms from "../../pages/legal-documents/terms-of-service/Terms-Of-Service.jsx";
-import NotFound from "../../pages/Error/400-error/NotFound.jsx";
-import Dashboard from "../../pages/inventory-read/Dashboard.jsx"
+import Home from "../../2-pages/home/Home.jsx";
+import PrivacyPolicy from "../../2-pages/legal-documents/privacy-policy/Privacy-Policy.jsx";
+import Terms from "../../2-pages/legal-documents/terms-of-service/Terms-Of-Service.jsx";
+import NotFound from "../../2-pages/Error/400-error/NotFound.jsx";
+import Dashboard from "../../2-pages/inventory-read/Dashboard.jsx"
 
 export default function Router() {
   return (
