+++ frontend/src/2-pages/inventory-read/Dashboard.jsx
