+++ frontend/src/2-pages/legal-documents/privacy-policy/Privacy-Policy.jsx
