+++ frontend/src/3-widgets/inventory/InventoryGrid.jsx
