+++ frontend/src/5-entities/inventory/inventoryApi.js
