+++ frontend/src/6-shared/components/Footer.jsx
