+++ frontend/src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import "./app/styles/index.css";
-import App from "./app/App.jsx";
+import "./1-app/styles/index.css";
+import App from "./1-app/App.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
